hwservicemanager: remove unsafe casts
Version x.(y+1) of a service must inherit from version x.y, so the
information of supported versions is already contained in the interface
chain. Because of this, we no longer need to parse interface names, we
can store interfaces using the version as a key, and we no longer need a
multimap.

Bug: 34085897
Test: hidl_test

Change-Id: I6a2ed9899ef1cb9a7ea37b6a3e8092fb9f09544d