Flush commands to binder driver.
To make sure it knows we've created a looper
for handling commands. This only becomes
necessary with new kernel changes that schedule
binder process operations to specific threads:
that code checks whether a thread has registered
before sending it operations, and without this
flush(), hwservicemanager's only thread won't be
marked as such.

Test: marlin boots
Change-Id: I38414cee4a6508148adf47c856524336b820edb1